+++ backend/application/socket/Impl/validacao_emit.py
@@ -58,20 +58,3 @@ def validacao_emit(json_emit: dict[str, any]):
         "Valido": len(erros) == 0,
         "Invalido": erros
     }
-
-def disparar_emit(socketio: SocketIO, evento: str, payload: dict, room: str | None = None):
-    try:
-        if not isinstance(payload, dict):
-            print(f"[Socket Error] payload não é um dicionario")
-            return
-        
-        payload_seguro = { **payload, "timestamp": datetime.now().isoformat() }
-        
-        if room: 
-            socketio.emit(evento, payload_seguro, room=room)
-        else:
-            socketio.emit(evento, payload_seguro)
-
-        print(f"[Socket Emit] evento: {evento} | room: {room}")
-    except Exception as e:
-       print(f"[Socket Error] falha ao emitir '{evento}': {str(e)}")
